update code, add fastapi_model
